Updated packaging for electron
